+++ src/utils.ts
@@ -171,7 +171,7 @@ function getRandomValuesWithMathRandom(bytes: Uint8Array): void {
 declare const msCrypto: Crypto;
 
 const getRandomValues = (() => {
-  if (typeof crypto !== 'undefined') {
+  if (typeof crypto !== 'undefined' && typeof crypto?.getRandomValues !== 'undefined') {
     return crypto.getRandomValues.bind(crypto);
   } else if (typeof msCrypto !== 'undefined') {
     return msCrypto.getRandomValues.bind(msCrypto);